Update documentation and personal portfolio
