+++ docs/UserGuide.adoc
@@ -179,7 +179,7 @@ Format: `note [PERSON] [STRING]`
 Examples:
 * `note John Doe Remember to call him once every day otherwise he will question your commitment.`
 
-
+//tag::photo[]
 === Attaching photos to persons :
 
 `Since v1.3` +
@@ -214,9 +214,16 @@ Restrictions on input:
 
 * Through Graphic User Interface
 
-In the `clients's detail` panel, user can use the `import photo` button to
+1. In the `person detail` panel, user can use the `import photo` button to
 add/edit photo of the client. Upon clicking the button, a file-chooser window
  will pop up, enabling the easy choosing of image files with GUI.
+2. A file-chooser window can also be opened with the command: `updatephoto`
+
+** Example command: `updatephoto 1` -> this will open a file-chooser window
+for the user to update photo of the person identified by index `1` in the
+last shown list.
+
+//end::photo[]
 
 === Setting relationship between 2 persons: `set`
 
@@ -345,7 +352,7 @@ Returns any event having titles that contain `Birthday`
 ====
 Pressing the kbd:[&uarr;] and kbd:[&darr;] arrows will display the previous and next input respectively in the command box.
 ====
-
+// tag::schedule[]
 === Displaying user's schedule: `checkschedule`
 
 `Since v1.4` +
@@ -375,6 +382,26 @@ is already occupied by an event, cannot drop at this time slot) +
 (eg. cannot drop at the last time slot, if event's duration is longer than 30
 minutes) +
 ****
+
+// end::schedule[]
+
+=== Schedule recurring event through `recurrence period` field
+
+User can schedule events with a regular period of recurrence through the
+`recurrence period` field, identified by prefix `re/` when adding or editing
+events. Values put for `recurrence period` can only be positive integers
+indicating interval between two scheduled occurrence of the event.
+
+For example, if an event is added with a  `repeat period` of 7 days, when the
+ event is finished, a new instance of the event will be scheduled with the same
+ time slot 7 days later. If this instance will clash with an existing event, it
+ will be skipped and a new instance will be created for the next period.
+
+When an event is added without specifying the `recurrence period`, the
+default vlaue is 0, which means it will not be repeated. Event recurrence can
+ also be removed by setting the value of the period to be 0.
+
+
 
 === Listing entered commands : `history`
