+++ src/main/java/seedu/address/model/event/exceptions/DuplicateEventException.java
@@ -2,7 +2,7 @@ package seedu.address.model.event.exceptions;
 
 import seedu.address.commons.exceptions.DuplicateDataException;
 
-// @@author yangshuang
+// @@author shuang-yang
 
 /**
  * Signals that the operation will result in duplicate Event objects.
