+++ src/main/java/seedu/address/ui/EventCard.java
@@ -8,7 +8,7 @@ import javafx.scene.layout.Region;
 import seedu.address.commons.core.index.Index;
 import seedu.address.model.event.ReadOnlyEvent;
 
-//@@author a0107442n
+//@@author shuang-yang
 
 /**
  * An UI component that displays information of a {@code Event}.
