+++ src/main/java/seedu/address/ui/PersonPanel.java
@@ -83,7 +83,7 @@ public class PersonPanel extends UiPart<Region> {
         registerAsAnEventHandler(this);
     }
 
-    //@@author a0107442n
+    //@@author shuang-yang
     /**
      * Register the image import button for click event.
      */
@@ -142,7 +142,7 @@ public class PersonPanel extends UiPart<Region> {
         tagsPane.getChildren().removeAll(tagsPane.getChildren());
         person.getTags().forEach(tag -> tagsPane.getChildren().add(new Label(tag.tagName)));
 
-        //@@author a0107442n
+        //@@author shuang-yang
         //Load the photo of the contact
         String imagePath = person.getPhoto().toString();
         Image image = new Image(new File(imagePath).toURI().toString());
