+++ src/main/resources/view/BrightTheme.css
@@ -1,4 +1,4 @@
-/* @@author a0107442n */
+/* @@author shuang-yang */
 
 .background {
     -fx-background-color: derive(white, 100%);
