+++ src/main/resources/view/EventListPanel.fxml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<!-- @@author a0107442n -->
+<!-- @@author shuang-yang -->
 
 <?import javafx.scene.control.ListView?>
 <?import javafx.scene.layout.VBox?>
