+++ src/main/resources/view/MainWindow.fxml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<!-- @@author yangshuang -->
+<!-- @@author shuang-yang -->
 
 <?import java.net.URL?>
 <?import javafx.geometry.Insets?>
