+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -7,6 +7,7 @@ import static org.junit.Assert.fail;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Timer;
 import java.util.function.Predicate;
 
 import org.junit.Rule;
@@ -21,6 +22,7 @@ import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.event.ReadOnlyEvent;
+import seedu.address.model.event.RepeatEventTimerTask;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.model.person.exceptions.DuplicatePersonException;
@@ -170,6 +172,11 @@ public class AddCommandTest {
         public ObservableList<ReadOnlyEvent> getSchedule() {
             fail("This method should not be called.");
             return null;
+        }
+
+        @Override
+        public void scheduleRepeatedEvent(ReadOnlyEvent addedEvent) {
+            fail("This method should not be called.");
         }
     }
