+++ src/test/java/seedu/address/model/AddressBookTest.java
@@ -118,6 +118,11 @@ public class AddressBookTest {
         }
 
         @Override
+        public ReadOnlyEvent getNewlyAddedEvent() {
+            return null;
+        }
+
+        @Override
         public ObservableList<ReadOnlyEvent> getSchedule(Date date) {
             return null;
         }
